# Documentation/openapi.yaml
openapi: 3.0.3
info:
  title: AIS Clinic API
  version: 1.0.0
  description: OpenAPI-описание ключевых эндпоинтов АИС "Клиника"
servers:
  - url: https://api.clinic.local
    description: Production
  - url: http://localhost:8000
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/visits:
    post:
      tags: [Visits]
      summary: Создать визит по записи
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VisitCreateRequest'
      responses:
        '201':
          description: Визит создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VisitResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
    get:
      tags: [Visits]
      summary: Список визитов с фильтрами
      parameters:
        - in: query
          name: doctor_id
          schema:
            type: string
            format: uuid
        - in: query
          name: patient_id
          schema:
            type: string
            format: uuid
        - in: query
          name: status
          schema:
            type: string
            enum: [in_progress, completed, cancelled]
        - in: query
          name: skip
          schema:
            type: integer
            minimum: 0
            default: 0
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
      responses:
        '200':
          description: Список визитов
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VisitListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/visits/{visit_id}:
    get:
      tags: [Visits]
      summary: Получить визит по id
      parameters:
        - in: path
          name: visit_id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Визит
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VisitResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/visits/{visit_id}/complete:
    patch:
      tags: [Visits]
      summary: Завершить визит
      parameters:
        - in: path
          name: visit_id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VisitCompleteRequest'
      responses:
        '200':
          description: Визит завершен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VisitResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
  /api/v1/visits/{visit_id}/diagnoses:
    post:
      tags: [Visits]
      summary: Добавить диагнозы к визиту
      description: |
        RBAC-правила:
        - admin может назначать диагнозы любому визиту в любом статусе;
        - doctor может назначать диагнозы только визитам своих пациентов и только после завершения приема (`appointment=completed` и `visit=completed`).
      parameters:
        - in: path
          name: visit_id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DiagnosesCreateRequest'
      responses:
        '201':
          description: Диагнозы добавлены
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DiagnosesCreateResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
  /api/v1/visits/{visit_id}/prescriptions:
    post:
      tags: [Visits]
      summary: Добавить назначения к визиту
      parameters:
        - in: path
          name: visit_id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PrescriptionsCreateRequest'
      responses:
        '201':
          description: Назначения добавлены
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericCreateResult'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
  /api/v1/visits/{visit_id}/referrals:
    post:
      tags: [Visits]
      summary: Добавить направления к визиту
      parameters:
        - in: path
          name: visit_id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReferralsCreateRequest'
      responses:
        '201':
          description: Направления добавлены
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReferralsCreateResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    ValidationError:
      description: Ошибка валидации входных данных
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Требуется аутентификация
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Недостаточно прав
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Ресурс не найден
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Конфликт состояния
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    VisitCreateRequest:
      type: object
      required: [appointment_id, chief_complaint, exam_notes]
      properties:
        appointment_id:
          type: string
          format: uuid
        chief_complaint:
          type: string
          minLength: 1
          maxLength: 4000
        exam_notes:
          type: string
          minLength: 1
          maxLength: 10000
        treatment_plan:
          type: string
          maxLength: 10000
    VisitCompleteRequest:
      type: object
      properties:
        treatment_plan:
          type: string
          maxLength: 10000
    VisitResponse:
      type: object
      required:
        [
          id,
          appointment_id,
          patient_id,
          patient_full_name,
          doctor_id,
          doctor_full_name,
          chief_complaint,
          exam_notes,
          visit_status
        ]
      properties:
        id:
          type: string
          format: uuid
        appointment_id:
          type: string
          format: uuid
        patient_id:
          type: string
          format: uuid
        patient_full_name:
          type: string
        doctor_id:
          type: string
          format: uuid
        doctor_full_name:
          type: string
        chief_complaint:
          type: string
        exam_notes:
          type: string
        treatment_plan:
          type: string
          nullable: true
        visit_status:
          type: string
          enum: [in_progress, completed, cancelled]
        closed_at:
          type: string
          format: date-time
          nullable: true
    VisitListResponse:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/VisitResponse'
        total:
          type: integer
          minimum: 0
    DiagnosisItem:
      type: object
      required: [icd10_code, is_primary]
      properties:
        icd10_code:
          type: string
          pattern: '^[A-Z][0-9]{2}(\.[0-9A-Z]{1,2})?$'
        is_primary:
          type: boolean
        comment:
          type: string
          maxLength: 2000
    DiagnosesCreateRequest:
      type: object
      required: [items]
      properties:
        items:
          type: array
          minItems: 1
          maxItems: 50
          items:
            $ref: '#/components/schemas/DiagnosisItem'
    DiagnosisResponse:
      type: object
      required: [id, icd10_code, is_primary]
      properties:
        id:
          type: string
          format: uuid
        icd10_code:
          type: string
          pattern: '^[A-Z][0-9]{2}(\.[0-9A-Z]{1,2})?$'
        is_primary:
          type: boolean
        comment:
          type: string
          maxLength: 2000
          nullable: true
    DiagnosesCreateResponse:
      type: object
      required: [visit_id, created, diagnoses]
      properties:
        visit_id:
          type: string
          format: uuid
        created:
          type: integer
          minimum: 0
        diagnoses:
          type: array
          items:
            $ref: '#/components/schemas/DiagnosisResponse'
    PrescriptionItem:
      type: object
      required: [medication_id, dosage, frequency, duration_days]
      properties:
        medication_id:
          type: string
          format: uuid
        dosage:
          type: string
          pattern: '^\d+(\.\d+)?\s?(mg|ml|g|mcg)$'
        frequency:
          type: string
          maxLength: 100
        duration_days:
          type: integer
          minimum: 1
          maximum: 365
        instructions:
          type: string
          maxLength: 2000
    PrescriptionsCreateRequest:
      type: object
      required: [items]
      properties:
        items:
          type: array
          minItems: 1
          maxItems: 50
          items:
            $ref: '#/components/schemas/PrescriptionItem'
    GenericCreateResult:
      type: object
      required: [visit_id, created]
      properties:
        visit_id:
          type: string
          format: uuid
        created:
          type: integer
          minimum: 0
    ReferralCreateItem:
      type: object
      required: [target_specialty_id, reason]
      properties:
        target_specialty_id:
          type: integer
          minimum: 1
        target_doctor_id:
          type: string
          format: uuid
          nullable: true
        reason:
          type: string
          minLength: 1
          maxLength: 2000
        comment:
          type: string
          maxLength: 2000
          nullable: true
    ReferralsCreateRequest:
      type: object
      required: [items]
      properties:
        items:
          type: array
          minItems: 1
          maxItems: 50
          items:
            $ref: '#/components/schemas/ReferralCreateItem'
    ReferralResponse:
      type: object
      required:
        [
          id,
          target_specialty_id,
          reason,
          status,
          created_at
        ]
      properties:
        id:
          type: string
          format: uuid
        target_specialty_id:
          type: integer
          minimum: 1
        target_doctor_id:
          type: string
          format: uuid
          nullable: true
        reason:
          type: string
        status:
          type: string
          enum: [created, accepted, rejected, completed]
        comment:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
    ReferralsCreateResponse:
      type: object
      required: [visit_id, created, referrals]
      properties:
        visit_id:
          type: string
          format: uuid
        created:
          type: integer
          minimum: 0
        referrals:
          type: array
          items:
            $ref: '#/components/schemas/ReferralResponse'
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, request_id]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  issue:
                    type: string
            request_id:
              type: string
              format: uuid